steppeau: Change the ethavb property as persistent property
To avoid auto launch of gptp process from init script,
setting the property to 0 would be required. Since this
property is not a persistent, on reboot property will not
retain the run time values. Make this property a persistent
property to retain its settings and set 0 as the its default
value.

Change-Id: I86ee19e809b05e8b8f6c44d432190dd9412eada2
Signed-off-by: Ramaprasad Nagaraj <ramapras@codeaurora.org>